requirements.txt
fastapi[all]==0.108.0
python-dotenv==1.0.0
yt-dlp==2024.03.10
redis==5.0.1
pymongo==4.6.1
pytz==2023.3.post1
aiohttp==3.9.1
